+++ .circleci/config.yml
@@ -33,6 +33,8 @@ jobs:
           # fallback to using the latest cache if no exact match is found
           - v1-dependencies-
 
+      - run: docker-compose up --build -d
+
       - run: mvn dependency:go-offline
 
       - save_cache:
