+++ data/db/WiredTiger
@@ -1,0 +1,2 @@
+WiredTiger
+WiredTiger 3.1.0: (April 23, 2018)
