+++ data/db/WiredTiger.lock
@@ -1,0 +1,1 @@
+WiredTiger lock file
