+++ docker-compose.yml
@@ -47,6 +47,8 @@ services:
     image: wurstmeister/zookeeper
     ports:
       - "2181:2181"
+    network_mode: "host"
+    hostname: localhost
     restart: always
     healthcheck:
       test: "exit 0"
@@ -56,11 +58,13 @@ services:
     ports:
       - "9092:9092"
     environment:
-      KAFKA_ADVERTISED_HOST_NAME: 192.168.99.100
-      KAFKA_ZOOKEEPER_CONNECT: zookeeper:2181
+      KAFKA_ADVERTISED_LISTENERS: PLAINTEXT://172.23.238.170:9092
+      KAFKA_LISTENERS: PLAINTEXT://0.0.0.0:9092
+      KAFKA_ZOOKEEPER_CONNECT: 172.23.238.170:2181
+    network_mode: "host"
+    hostname: localhost
     depends_on:
-      zookeeper:
-        condition: service_healthy
+      - zookeeper
     restart: always
     healthcheck:
       test: "exit 0"
@@ -94,10 +98,6 @@ services:
         condition: service_healthy
       mongodb:
         condition: service_healthy
-      zookeeper:
-        condition: service_healthy
-      kafka:
-        condition: service_healthy
     healthcheck:
       test: "exit 0"
     
@@ -117,10 +117,6 @@ services:
       eureka-server:
         condition: service_healthy
       mongodb:
-        condition: service_healthy
-      zookeeper:
-        condition: service_healthy
-      kafka:
         condition: service_healthy
     healthcheck:
       test: "exit 0"
