--- upstream-service/src/main/java/com/stackroute/eplay/streams/RSVPEventStreams.java
@@ -1,12 +1,0 @@
-package com.stackroute.eplay.streams;
-
-import org.springframework.cloud.stream.annotation.Output;
-import org.springframework.messaging.MessageChannel;
-
-//Stream defined for writing messages to Message Bus
-public interface RSVPEventStreams {
-
-    String OUTPUT = "RSVP-event-out";
-    @Output(OUTPUT)
-    MessageChannel outboundEvents();	// to write messages to kafka
-}
