+++ upstream-service/src/main/java/com/stackroute/eplay/upstreamservice/controller/UpStreamController.java
@@ -8,12 +8,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.stackroute.eplay.streams.MovieEventStreams;
-import com.stackroute.eplay.streams.MovieStreams;
-import com.stackroute.eplay.streams.RSVPEventStreams;
-import com.stackroute.eplay.streams.ShowStreams;
-import com.stackroute.eplay.streams.TheatreStreams;
-import com.stackroute.eplay.streams.TicketedEventStreams;
 import com.stackroute.eplay.upstreamservice.domain.Movie;
 import com.stackroute.eplay.upstreamservice.domain.MovieEvent;
 import com.stackroute.eplay.upstreamservice.domain.RSVPEvent;
@@ -21,6 +15,12 @@ import com.stackroute.eplay.upstreamservice.domain.Show;
 import com.stackroute.eplay.upstreamservice.domain.Theatre;
 import com.stackroute.eplay.upstreamservice.domain.TicketedEvent;
 import com.stackroute.eplay.upstreamservice.service.UpStreamService;
+import com.stackroute.eplay.upstreamservice.streams.MovieEventStreams;
+import com.stackroute.eplay.upstreamservice.streams.MovieStreams;
+import com.stackroute.eplay.upstreamservice.streams.RSVPEventStreams;
+import com.stackroute.eplay.upstreamservice.streams.ShowStreams;
+import com.stackroute.eplay.upstreamservice.streams.TheatreStreams;
+import com.stackroute.eplay.upstreamservice.streams.TicketedEventStreams;
 
 @RestController
 @EnableBinding({TicketedEventStreams.class,MovieEventStreams.class,RSVPEventStreams.class,TheatreStreams.class,MovieStreams.class, ShowStreams.class})
